Getting a better handle on things..  offtree_region_preprw() will contain the entire mechanics for growing the tree, placing memory allocations, and collecting the array of iovec's needed to map the given region. struct offtree_req holds the dynarray which, when complete, will hold those iovecs.
Still need to do the part where the tree children and allocated and assigned their buffers, 
this will have to be recursive since a single allocation may require multiple depth traversals 
for proper accommodation of the buffers.

Also the race conditions need to be reconsidered... I presume this will be very difficult.  
One thing is certain when a tree leaf is splitting or no more allocations are allowed until 
offtree_region_preprw() releases the respective nodes.  This way we won't have multiple 
buffers for the same off+len pairs.



git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@2701 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda