when completing filling of a dircache page, do not block if we find a page in the loading state.
we were doing this to prevent the same d_off from being in two pages.
without this safety, we may run into problems, but the current situation
is worse.

as this blocking happens in RPC callback context, multiple threads preloading
pages simultaneously currently quickly ties up all RPC handling threads
and pages in the queue that need to be processed to wake all stuck RPC
threads will not be processed as the threads are all stuck!


git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@23000 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda