generalize predictive I/O to accomodate upcoming bmap "writeahead"
why?  in certain circumstances, it makes sense to preload a write lease
for anticipated write I/O that extends into the next bmap.

the worry is that, with multiple writers, that will will degrade to DIO
when this writer has no actual intention of writing into the bmap.

in these situations, extra RPCs to relinquish the "advisory writeahead
lease" will be needed and will impose RPC round trip penalty on the
action.

otherwise,
  - for single stream writing apps, performance should improve greatly
  - for multiple writers that are not specifically catering to bmap
    boundaries, advisory leases are ignored by the MDS if already bound,
    so real effects should be negligible.  and we can always add a
    tunable to outright disable the feature.


git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@25074 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda