When doing slash2 self build, I see generation number going backwards in a setattr code path. Twice they happen on a directory.  As it turns out, we use generation number to track directory modifications as well.
The code path involved is a readdir callback, which should actually only updates attributes when
the fcmh is not already there. If not, the use of SAVE_LOCAL flag is totally bogus because we
have no attributes to start with.

This patch fixes the above issues.  I was able to run four way tests (self build, kernel build,
iozone, tar over ssh) for three hours without seeing any problems.  Yay!