If we encounter an error (e.g., EKERYEXPIRED), the sliver won't be able to serve I/O for a loooong time. We want to let these slivers to die so we can recover from transient errors.
The patch does that. For this to be safe, we must reclaim bad slivers immediately.  I did fault inject with
the following command:

./slictl -p faults.fsio_read_fail.active=1

and later clear it.  Client is able to read the file again.  I also did self build and iozone -a. They seems
to be working.

This is a risky change.  It requires no one touch the sliver after read/write io_done() function, which seems
to be the case.  bascially, sliver can be freed after you drop the reference.