fix some fallout from the restructuring to support asynchronous bmap retrieval:  the error code returned by the BMAP_LEASE_GET RPC is ignored in the blocking case and always returns success, which copies a zeroed bmapdesc that is used later.
sanity checks on the bmapdesc were added to catch this class of error in
the future.

this fixes a very simple multi-client error as seen on multiple SLASH2
deployments recently:

 hostA$ echo test > test

 hostB$ ls test
 test
 hostB$ cat test
 cat: No such file or directory

ENOENT was actually being returned to read(2) because a tryext() was
being called on a zereod lease since the bmap needs to DIOWAIT before
the IO can succeed.