cleanup fidc_lookup APIs
this API has devolved into an undisciplined mess over the years with different
naming conventions and inconsistent behaviors, and lots of hand-rolling.

the new convention is simply:

- fidc_lookup()
	this is the master call that takes all arguments, that you rarely want
	to use.

- sl_fcmh_*()
	these are the subset of routines used in 95% of all places.  exceptions
	use long-rolled invocations with default parameter values.

to minimize this diff, a small subset of per-daemon routines (slm_fcmh_*,
sli_fcmh_*) are preserved.

underscore versions of all routines are simply pfl_callerinfo versions of the
same routine.