revise the operation of the dirent namecache.
previously, the namecache was laden with race conditions, especially
when async operations (READDIR reply handling) are involved.

the new model switches to windows when the namecache may change by
marking dirents in the namecache as HOLD, meaning they cannot be
accessed.  during such times, accesses fail and go to the authoritative
MDS.