Don't hold fcmh lock after mark it as busy.
This is a risky change.  The idea is to fix deadlock seen with replication
add/removal testing with a file that has many bmaps.  Basically, if we
hold fcmh lock and then wait for the bmap's flag to clear, the one and only
work thread won't be remove the bmap.  The last step of removing a bmap
involving taking a fcmh lock (see bmap_remove().