I hit a crash on FREEING bit because the same bq is called twice: once with -110 and the other with -104.
It turns out that using worker thread enlarge the race window, make
it happen more likely.  In this patch, I remove the bq from the wait
list before scheduling a done.  This should be better, but might not
fix the issue depending on the LNET layer: when it calls individual
callback and when it calls the drop connection callback.