Complains from a user and a fix:
I just confirmed that a user who is in a group with ACLs that set read-only access to a directory is then
able to go and set rwx permissions on that directory for another group. I tried to duplicate this on /pylon1
and it did not allow the user with read permission to set the ACLs on a directory.